Add map layers back to App while keeping map stuff in Map, add tooltip to stops.
