+++ src/components/App.js
@@ -10,17 +10,51 @@ import {inject, observer} from "mobx-react";
 import Map from "./map/Map";
 import {latLng} from "leaflet";
 import getCoarsePositionForTime from "../helpers/getCoarsePositionForTime";
+import map from "lodash/map";
+import StopLayer from "./map/StopLayer";
+import RouteQuery from "../queries/RouteQuery";
+import RouteLayer from "./map/RouteLayer";
+import HfpLayer from "./map/HfpLayer";
+import HfpMarkerLayer from "./map/HfpMarkerLayer";
+import invoke from "lodash/invoke";
 
 @inject(app("Journey"))
 @withHfpData
 @observer
 class App extends Component {
-  render() {
+  state = {
+    stopsBbox: null,
+  };
+
+  onMapChanged = (map) => {
+    this.setStopsBbox(map);
+  };
+
+  setStopsBbox = (map) => {
+    if (!map) {
+      return;
+    }
+
+    const bounds = map.getBounds();
+
+    if (!bounds || !invoke(bounds, "isValid")) {
+      return;
+    }
+
+    this.setState({
+      bbox: {
+        minLat: bounds.getSouth(),
+        minLon: bounds.getWest(),
+        maxLat: bounds.getNorth(),
+        maxLon: bounds.getEast(),
+      },
+    });
+  };
+
+  getJourneyBounds = () => {
     const {
-      loading,
       state: {selectedJourney, date, time},
       positionsByJourney,
-      positionsByVehicle,
     } = this.props;
 
     let journeyBounds = null;
@@ -40,10 +74,64 @@ class App extends Component {
       }
     }
 
+    return journeyBounds;
+  };
+
+  render() {
+    const {stopsBbox} = this.state;
+    const {loading, state, positionsByVehicle, Journey} = this.props;
+    const {route, vehicle, stop, selectedJourney} = state;
+
+    const journeyBounds = this.getJourneyBounds();
+
     return (
       <div className="transitlog">
         <FilterPanel />
-        <Map positionsByVehicle={positionsByVehicle} bounds={journeyBounds} />
+        <Map onMapChanged={this.onMapChanged} bounds={journeyBounds}>
+          {!route &&
+            map.zoom > 14 && <StopLayer selectedStop={stop} bounds={stopsBbox} />}
+          <RouteQuery route={route}>
+            {({routePositions, stops}) => (
+              <RouteLayer
+                route={route}
+                routePositions={routePositions}
+                stops={stops}
+                setMapBounds={this.setMapBounds}
+                key={`route_line_${route}`}
+                hfpPositions={positionsByVehicle}
+              />
+            )}
+          </RouteQuery>
+          {positionsByVehicle.length > 0 &&
+            positionsByVehicle.map(({positions, vehicleId}) => {
+              if (vehicle && vehicleId !== vehicle) {
+                return null;
+              }
+
+              const lineVehicleId = get(selectedJourney, "uniqueVehicleId", "");
+              const journeyStartTime = get(selectedJourney, "journeyStartTime", "");
+
+              const key = `${lineVehicleId}_${route}_${journeyStartTime}`;
+
+              return [
+                lineVehicleId === vehicleId ? (
+                  <HfpLayer
+                    key={`hfp_line_${key}`}
+                    selectedJourney={selectedJourney}
+                    positions={positions}
+                    name={vehicleId}
+                  />
+                ) : null,
+                <HfpMarkerLayer
+                  key={`hfp_markers_${route}_${vehicleId}`}
+                  onMarkerClick={Journey.setSelectedJourney}
+                  selectedJourney={selectedJourney}
+                  positions={positions}
+                  name={vehicleId}
+                />,
+              ];
+            })}
+        </Map>
         <LoadingOverlay show={loading} message="Ladataan HFP-tietoja..." />
       </div>
     );
