+++ src/components/map/Map.js
@@ -1,15 +1,8 @@
 import React, {Component} from "react";
 import {observer, inject} from "mobx-react";
-import StopLayer from "./StopLayer";
-import RouteQuery from "../../queries/RouteQuery";
-import RouteLayer from "./RouteLayer";
-import get from "lodash/get";
-import HfpLayer from "./HfpLayer";
-import HfpMarkerLayer from "./HfpMarkerLayer";
 import {LeafletMap} from "./LeafletMap";
 import {app} from "mobx-app";
 import invoke from "lodash/invoke";
-import map from "lodash/map";
 
 @inject(app("Journey"))
 @observer
@@ -21,7 +14,6 @@ class Map extends Component {
   };
 
   state = {
-    bbox: null,
     bounds: null,
     lat: 60.170988,
     lng: 24.940842,
@@ -35,7 +27,6 @@ class Map extends Component {
   };
 
   onMapChanged = (map, viewport) => {
-    this.getBbox(map);
     this.props.onMapChanged(map, viewport);
   };
 
@@ -49,31 +40,9 @@ class Map extends Component {
     this.props.onMapChange(map, viewport);
   };
 
-  getBbox = (map) => {
-    if (!map) {
-      return;
-    }
-
-    const bounds = map.getBounds();
-
-    if (!bounds || !invoke(bounds, "isValid")) {
-      return;
-    }
-
-    this.setState({
-      bbox: {
-        minLat: bounds.getSouth(),
-        minLon: bounds.getWest(),
-        maxLat: bounds.getNorth(),
-        maxLon: bounds.getEast(),
-      },
-    });
-  };
-
   render() {
-    const {positionsByVehicle, state, Journey, bounds: propBounds} = this.props;
-    const {bbox, lat, lng, zoom, bounds} = this.state;
-    const {route, vehicle, stop, selectedJourney} = state;
+    const {bounds: propBounds, children} = this.props;
+    const {lat, lng, zoom, bounds} = this.state;
 
     const useBounds = propBounds || bounds || null;
 
@@ -84,48 +53,7 @@ class Map extends Component {
         bounds={useBounds}
         onMapChanged={this.onMapChanged}
         onMapChange={this.onMapChange}>
-        {!route && map.zoom > 14 && <StopLayer selectedStop={stop} bounds={bbox} />}
-        <RouteQuery route={route}>
-          {({routePositions, stops}) => (
-            <RouteLayer
-              route={route}
-              routePositions={routePositions}
-              stops={stops}
-              setMapBounds={this.setMapBounds}
-              key={`route_line_${route}`}
-              hfpPositions={positionsByVehicle}
-            />
-          )}
-        </RouteQuery>
-        {positionsByVehicle.length > 0 &&
-          positionsByVehicle.map(({positions, vehicleId}) => {
-            if (vehicle && vehicleId !== vehicle) {
-              return null;
-            }
-
-            const lineVehicleId = get(selectedJourney, "uniqueVehicleId", "");
-            const journeyStartTime = get(selectedJourney, "journeyStartTime", "");
-
-            const key = `${lineVehicleId}_${route}_${journeyStartTime}`;
-
-            return [
-              lineVehicleId === vehicleId ? (
-                <HfpLayer
-                  key={`hfp_line_${key}`}
-                  selectedJourney={selectedJourney}
-                  positions={positions}
-                  name={vehicleId}
-                />
-              ) : null,
-              <HfpMarkerLayer
-                key={`hfp_markers_${route}_${vehicleId}`}
-                onMarkerClick={Journey.setSelectedJourney}
-                selectedJourney={selectedJourney}
-                positions={positions}
-                name={vehicleId}
-              />,
-            ];
-          })}
+        {children}
       </LeafletMap>
     );
   }
